requirements.txt
streamlit
pandas
sklearn-pandas